+++ noodles-cram/src/lib.rs
@@ -11,6 +11,7 @@ pub mod crai;
 pub mod data_container;
 pub mod file_definition;
 mod huffman;
+mod indexed_reader;
 mod indexer;
 pub(crate) mod io;
 mod num;
@@ -19,8 +20,8 @@ pub mod record;
 pub mod writer;
 
 pub use self::{
-    data_container::DataContainer, file_definition::FileDefinition, indexer::index, reader::Reader,
-    record::Record, writer::Writer,
+    data_container::DataContainer, file_definition::FileDefinition, indexed_reader::IndexedReader,
+    indexer::index, reader::Reader, record::Record, writer::Writer,
 };
 
 #[cfg(feature = "async")]